Redirect assets if redirect_url is present before auth
This commit moves redirection of assets if a `redirect_url` is present to happen _before_ auth. The reasoning behind this commit comes from a recent issue where an unpublished document was linked to via origin, which was then redirected to the draft host, requiring the end user to authenticate. Had the end user successfully authenticated, they would then have proceeded to the draft asset host to be redirected back to origin on a different path.

By moving redirection of assets before authentication takes place, requests will get to their final destination first _before_ authentication is performed, if necessary.